+++ arcane/src/arcane/utils/srcs.cmake
@@ -27,7 +27,6 @@ set(ARCANE_SOURCES
   CommandLineArguments.cc
   Convert.cc
   Convert.h
-  ConcurrencyUtils.cc
   ConcurrencyUtils.h
   CStringUtils.cc
   CStringUtils.h
@@ -354,5 +353,4 @@ set(ARCANE_SOURCES
   internal/MemoryBuffer.h
   internal/MemoryPool.h
   internal/ParallelLoopOptionsProperties.h
-  internal/TaskFactoryInternal.h
   )
